moved code from main into separate files
